pyproject.toml
[project]
name = "mcp-patterns"
version = "0.1.0"
description = "Reusable patterns for building MCP servers"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0,<3",
    "httpx>=0.27",
    "pydantic>=2.0",
    "PyJWT>=2.8",
    "langchain-mcp-adapters>=0.1",
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "chainlit>=2.0",
    "python-dotenv>=1.0",
    "uvicorn>=0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_patterns", "src/server", "src/client"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
